refactor(api): migrate campaign-participation-service
