+++ api/src/prescription/campaign/domain/services/campaign-participation-service.js
