+++ api/tests/prescription/campaign/unit/domain/services/campaign-participation-service_test.js
@@ -1,5 +1,5 @@
-import { expect } from '../../../test-helper.js';
-import * as campaignParticipationService from './../../../../lib/domain/services/campaign-participation-service.js';
+import * as campaignParticipationService from '../../../../../../src/prescription/campaign/domain/services/campaign-participation-service.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Service | Campaign Participation Service', function () {
   describe('progress', function () {
